Move TL zombifying logic to tl-corruption-detection project
