+++ timelock-agent/src/main/java/com/palantir/timelock/paxos/TimeLockAgent.java
@@ -44,9 +44,9 @@ import com.palantir.atlasdb.timelock.TooManyRequestsExceptionMapper;
 import com.palantir.atlasdb.timelock.adjudicate.FeedbackHandler;
 import com.palantir.atlasdb.timelock.adjudicate.HealthStatusReport;
 import com.palantir.atlasdb.timelock.adjudicate.TimeLockClientFeedbackResource;
-import com.palantir.atlasdb.timelock.corruption.CorruptionNotifierResource;
-import com.palantir.atlasdb.timelock.corruption.JerseyCorruptionFilter;
-import com.palantir.atlasdb.timelock.corruption.UndertowCorruptionHandlerService;
+import com.palantir.timelock.corruption.detection.CorruptionNotifierResource;
+import com.palantir.timelock.corruption.detection.JerseyCorruptionFilter;
+import com.palantir.timelock.corruption.detection.UndertowCorruptionHandlerService;
 import com.palantir.atlasdb.timelock.lock.LockLog;
 import com.palantir.atlasdb.timelock.lock.v1.ConjureLockV1Resource;
 import com.palantir.atlasdb.timelock.management.PersistentNamespaceContext;
