+++ timelock-corruption-detection/src/main/java/com/palantir/timelock/corruption/detection/RemoteCorruptionDetector.java
@@ -14,8 +14,17 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.timelock.corruption;
+package com.palantir.timelock.corruption.detection;
 
-public interface CorruptionDetector {
-    boolean hasDetectedCorruption();
+public class RemoteCorruptionDetector implements CorruptionDetector {
+    private volatile CorruptionStatus remoteCorruptionState = CorruptionStatus.HEALTHY;
+
+    public void setRemoteCorruptionState() {
+        this.remoteCorruptionState = CorruptionStatus.CORRUPTION;
+    }
+
+    @Override
+    public boolean hasDetectedCorruption() {
+        return remoteCorruptionState.hasCorruption();
+    }
 }
